refactor: middlewares
