+++ middlewares/index.js
@@ -1,0 +1,6 @@
+module.exports.cors = require('./cors');
+module.exports.rateLimiter = require('./rateLimiter');
+module.exports.notFoundHandler = require('./notFoundHandler');
+module.exports.errorHandler = require('./errorHandler');
+module.exports.errorLogger = require('./errorLogger');
+module.exports.requestLogger = require('./requestLogger');
