+++ middlewares/notFoundHandler.js
@@ -1,3 +1,3 @@
-const NotFoundError = require('../errors/404');
+const NotFoundError = require('../utils/errors/404');
 
 module.exports = (req, res, next) => next(new NotFoundError('Запрашиваемый ресурс не найден.'));
